Fix handling of ready state.
Some TPMs implement an Idle state which is entered prior to the Ready state.  This requires an extra state transition in order to enter the Ready state.  Thus, tpm_wait_cmd_ready() now sets the command_ready bit within the loop that waits for the TPM to indicate that it is ready in order to ensure that any needed state transitions get made.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>
ACKed-by:  Gang Wei <gang.wei@intel.com>