Simplified tboot binary build process to use the standard ELF process and tools
This removes the Xen-derived, complex process used previously.  That process was intended for the case that the binary needed to load at a different address than it was compiled to run at--this is not true of tboot.  Thus, the tools in the build_tools dir are no longer needed and are removed.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>