On (re-)entry from SENTER, verify that physical page layout of MLE page table matches that of tboot
Per the MLE Developers Guide, on entry from SINIT the MLE must verify the physical address layout if is is not using paging.  This is to prevent malware from having moved the pages to different physical addresses (but not having changed the content or the relative order, and thus the hash being the same) and causing corruption of the execution flow (since the code will have the expected physical addresses in it).

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>
ACKed-by: Shane Wang <shane.wang@intel.com>