Expand VT-d protections to include all of usable RAM (previous was just tboot)
The VT-d PMRs will now be set to protect the following memory ranges:
RAM < 4GB:  0 to the last address marked as E820_RAM in the e820 table
RAM > 4GB:  the lowest address marked as E820_RAM to the highest address marked as E820_RAM

The upper/limit addresses will be rounded down to the nearest 2MB to avoid any overlap of neighboring reserved regions.  It has been found that some systems will hang if DMA to reserved regions is prevented.  Any delta due to rounding is marked as reserved in the e820 table to prevent its use.

The VT_D build option is removed.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>
ACKed-by: Shane Wang <shane.wang@intel.com>