Support new shutdown type, TB_SHUTDOWN_WFS, for APs to enter wait-for-SIPI state
This is needed to support Linux S4, where the kernel puts the APs into wfs before it generates the hibernation image and then wakes them before putting the platform into S4.

The AP mini-guests now spin on a memory flag to indicate if they should exit the guest before INIT-SIPI-SIPI.  This allows the system to be gracefully shutdown if an error is encountered in tboot after they have been awakened and before the OS has taken them out of wfs.  It is also need for Linux, which puts APs into wfs before platform Sx.

The tboot_shared struct version is incremented to 5.  A new field, num_in_wfs (and paddign to 32b -align it), was added so that the kernel/VMM can determine when all APs have entered/exited the mini-guests.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>