Fixed bug in tpm_seal() with setting of localityAtCreation and added fallback for tpm_get_random()
tpm_seal() was failing on Winbond/Nuvoton TPMs because localityAtCreation was being set to 0x1f. It is now set to the locality at which the command is being sent.

tpm_get_random() was failing to get all of the requested random bytes on a Winbond/Nuvoton TPM.  tpm_get_random() will now try a second time to get the remaining requested bytes.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>