Reserved tboot's memory as E820_RESERVED if launching Linux; fix tboot image size
If launching Linux, mark tboot's memory in the e820 table as E820_RESERVED instead of E820_UNUSABLE.  Xen is unchanged.  This is a more proper memory type as there is no guarantee that the kernel woudl maintain proper MTRR settings over an unusable region, etc.

Now that tboot DMA protects all of DRAM, the tboot reserved memory region only needs to be rounder up to a page boundary instead of a 2MB boundary.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>