Copy LCP owner policy data (if present) to buffer in os_mle_data
If an LCP owner policy data file was provided (i.e. loaded by GRUB), then that policy data will be needed for all launches of tboot--both initial and S3 resume.  So in order to ensure that it is available for S3 resume, we copy it to a (new) buffer in os_mle_data.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>