Support TPMs that return TPM_RETRY for TPM_SaveState command
Some TPMs may return TPM_RETRY for the TPM_SaveState command if the command is being executed a second time, in order to prevent NV wearout from repated calls.  In such cases, the TPM will allow the second invocation after ~5 sec.  This has been observed on an Infineon TPM.

The tpm_save_state() function will now retry the command as long as it returns TPM_RETRY, up to MAX_SAVESTATE_RETRIES times (currently defined to be 1000).  It will print a '.' for each retry, as a progress indicator.