Always extend PCRs on S3 integrity creation/verification
Regardless of whether S3 integrity creation (i.e. sealing of S3 data) or verification succeeds or fails, PCR(s) will always be extended.  In the case of creation failure, the PCRs will be extended with the measured values.  In the case of verification failure, we can't trust the values so we cap PCR 18.

If the PCR extend operation itself fails, this is a fatal error, as we cannot let execution continue without some distinguishing value being extended.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>