Fix bug where some e820 configurations would case recent USB-related fix code to reserve too much memory
The recent change to not DMA protect USB buffers has a bug where, for some e820 configurations, it will reserve most of the free memory as being above the last reserved region.  This change fixes that and optimizes the reserving of the VT-d truncation memory so that it doesn't re-walk the e820 table.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>