Fixed buffer overrun errors in lcptools/ and compilation error on some systems in utils/
In lcptools/, crtpol and crtpolelt had buffer overrun bugs that are now fixed.

parse_err.c and txt-stat.c assumed that sys/user.h and/or sys/mman.h would provide the definition of PAGE_SHIFT and PAGE_SIZE.  However, this is not the case on all systems.  Moreover, the tboot files that use these assume that they are 4k pages.  This change defines this explicitly to be 4k.

The error in crtpolelt.c was discovered by Martin Pirker.  The others were reported by Michael Gissing.  Fixes for all were provided by Michael Gissing.

Signed-off-by: Michael Gissing <m.gissing@tugraz.at>
Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>