Add timeout ot comc_setup() and don't re-initialize serial on shutdown
A timeout (the same value used for serial xmit) was added to comc_setup() to prevent an infinite loop shoudl the serial port never get into the ready state.

On shutdown (particularly S3), the OS/VMM may disable/power down the serial port.  This will cause a delay when tboot tries to re-initialize it.  So the re-initialization has been removed and tboot will simply use it with its current settings (or its use will fail if it was disabled/powered down).

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>
ACKed-by: Shane Wang <shane.wang@intel.com>