Removed no_usb command line parameter and SMI disabling
It was found that on some BIOSes and platforms, when legacy USB SMIs were disabled, that the system would reset or that it would generate DMA errors in an internal register.  Given this, it is best to remove this capability from the code and revert to the previous behavior of simply reserving any memory above the upper reserved regions and tuncating for VT-d.

This change essentially reverts c/s 217.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>