Separate AC Module subtype field from type field in ACM header
The existing code used a 32 bit type field in the ACM header.  But the type field is actually only 16 bits followed by a 16 bit subtype field.  The acmod.h file is changed appropraitely and the dump of the ACM header (in both tboot and acminfo output) is changed to show the subtype field.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>