TPM: add waiting on expect==0 before issue tpmGo
This is bug fix to keep complying with TCG PC Client TIS Spec which indicates that TPM will ignores the state when Expect (E) is 1 when tpmGO (TG) is 1.

Affected platform as a consequence to the wrong sequence during command transmission, the TPM correctly ignores the write to tpmGo as it is not yet in the state that all bytes have been received and waits forever for the tpmGo after the TPM has received all bytes.

Signed-off-by: Sevaiyan Balaguru <sevaiyan.balaguru@intel.com>
Signed-off-by: Gang Wei <gang.wei@intel.com>