Fix for possible overwritting to mle page table by GRUB2
MLE page table are 3 pages just below tboot start memory. It was not reserved via linker previously, so that when tboot is launched by GRUB2, GRUB2 may overwrite those 3 undeclared pages when it try to utilize the memory below tboot memory.

The fix is just define a 3-page arrary and put it at the start of .text section, meanwhile change the .text section start to TBOOT_BASE_ADDR to keep the _mle_start unchanged.

Signed-off-by: Gang Wei <gang.wei@intel.com>