There is a pointer dereference regression in the tboot native Linux loader which manifests itself as a system reset after the following is displayed on systems which are not using an initial ramdisk or initramfs image.
TBOOT: transferring control to kernel @0x100000

The native linux loader does not check the multiboot module count when
initializing the initrd/initramfs image size and pointer value.
This causes the loader setup code to pass an invalid pointer along with
an arbitrary size count to the loader which then attempts to copy the
contents of whatever memory the pointer is referencing into the kernel
image that is being built.

The fix is straight forward.  If the remaining multiboot module count is
zero the initrd image pointer and size value are explicitly set to zero.
This condition is interpreted properly by the loader as an indication that
an initrd image is not to be loaded.

Signed-off-by: Dr. Greg Wettstein <greg@enjellic.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>