Ensure tboot log is available even when measured launch is skipped.
Tboot only protects the log area in the e820 map in post_launch i.e. only if
a measured launch has successfully been performed. This means that in cases
where measured launch is skipped, tboot log’s memory is not reserved in the
e820 map and can be reused by the kernel. This can make it hard to figure out
why measured launch was skipped.

This change moves the code to reserve the tboot log in the E820 map to the
start of launch_kernel, which is called regardless of whether measured launch
succeeded or not.

Testing:  Disable TXT. Boot host. Verify no log is displayed without fix, and
tboot log is displayed with fix.



Signed-off-by: Sahil Rihan <srihan@fb.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>