Add support for appending to a TPM2 TCG style event log.
Tboot didn’t have support for appending to a TPM2 TCG style event log. As a result
it would crash with a “generic fatal error” when it ran with an SINIT that supports
the TCG log format.

This change addresses that by adding code to handle the new log format.

Testing done: Verify tboot crashes on Skylake system without patch, and successfully
boots kernel with the patch.

Signed-off-by: Sahil Rihan <srihan@fb.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>