Add an option in tboot to force SINIT to use the legacy TPM2 log format.
Some SINITs have a bug where they don't extend the MLE hash to the event log.
This makes it impossible to verify the measurement chain for PCR 17. However,
if we force them to use the legacy (not TCG standardized) TPM2 log format, the
SINITs in question log all the inputs to PCR 17 to the event log. This setting
provides a way to force use of the legacy log format for TPM2 systems.

Testing done:  Run tboot with and without option on TPM2 system that supports TCG
log format. Verify MLE hash entry (0x404) is missing from event log when system is
booted without the option (or option set to false). Verify legacy log format is used
and MLE hash entry is present when system is booted with option set to true.


Signed-off-by: Sahil Rihan <srihan@fb.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>