Enable VGA logging for EFI platforms
TBOOT uses old 80x25 BIOS console support for VGA logging. This method is not
supported by modern EFI platforms. This patch adds text render that prints
messages using framebuffer (GRUB2 is responsible to select proper fb mode and
pass information in multiboot2). When framebuffer is not available, 80x25
console is used as a fallback.

There is a limitation of that approach - it cannot be used when PMRs are
programmed. That's why, VGA logging is disabled after executing GETSEC[SENTER].

Signed-off-by: Lukasz Hawrylko <lukasz.hawrylko@intel.com>