Compress memory log before setting MTRRs
Memory logging mechanism collect messages to the buffer and if buffer grows above
threshold it is compressed. Compression should not be executed when MTRRs are
programmed for GETSEC[SENTER] as it will take very long time. To prevent that
situation, this patch forces to do the compression, even if threshold is not
reached, before programming MTRRs.

Signed-off-by: Lukasz Hawrylko <lukasz.hawrylko@intel.com>