make efi_memmap_reserve handle gaps like e820_protect_region already does
This revision corrects an inconsistency in the way efi_memmap_reserve
deals with requests that lie within a gap in the memory map.

In such cases, e820_protect_region() created a new range in the map,
however efi_memmap_reserve was doing nothing since its logic was
based on finding an overlapping range.

This revision makes efi_memmap_reserve add a new range.  Unlike
the e820 case, the EFI memory map have an "attributes" field as well
and it is somewhat ambigious how to define the attributes of a new map.
Since we are only using this for reserved ranges, the attributes are set to zero
for simplicity...

Signed-off-by: Alex Olson <alex.olson@starlab.io>