Ensure that growth of Multiboot tags does not go beyond original area
The grow_mb2_tag() function can only rearrange Multiboot content
within the area originally allocated by the parent bootloader (Grub).

This revision ensures that boot will not continue if the expansion of
a tag would overflow the original area.

In practice, small expansions as a result of tboot added content
are normally "paid for" through tboot's removal of other tags.

Signed-off-by: Alex Olson <alex.olson@starlab.io>