Replace EFI memory map in Multiboot2 info
tboot modifies both the e820 and EFI memory maps during its boot process,
but was only updating the Multiboot content's e820 map.  This meant
the unmodified EFI map was being passed even though it should have contained
reserved ranges specific to tboot.

This probably went unnoticied since the Linux kernel is not booted using
Multiboot (thus, was already receiving the updated EFI map on EFI systems).

This revision adds the update of the EFI Multiboot content, but has to make a
slight tradeoff as the combined expansion of the E820 and EFI memory maps
cannot be "paid" for with the memory savings from other Multiboot tag deletions.

Signed-off-by: Alex Olson <alex.olson@starlab.io>