Implement ACM flexible info table handling
ACM Header v5 introduces ACM Info Table v9 in which chipset ID,
processor ID, TPM info lists are moved from their original place
and there is a second chipset ID list added.

This patch adds handling for these lists.

(This is a first draft based on TXT MLE Dev Guide and is subject
to changes since it is not yet possible to fully test it. It does
not break current functionality.)