Fix the RAM memory allocation algorithm for the initrd.
It was noticed on some of the Intel boards, that TBOOT can't find
any available memory chunk in area and allocates the initial RAMDISK
in the kernel space memory. In order to resolve this issue, the improved algorithm
combines the RAM memory regions, unused by the Linux kernel(EFI_LOADER_CODE,
EFI_LOADER_DATA, EFI_CONVENTIONAL_MEMORY) and then searches for the biggest RAM memory
chunk for the initrd allocation. The given change allows in those cases
to allocate initrd, beyond the kernel memory range.