Increase the maximal APIC ID value from 1024 to 8192.
On the Intel Birchstream multisocket server platforms, it was noticed that after GETSEC[WAKEUP]
command execution, TBOOT was hanging, during the RLPs wakeup process. This problem has been caused
by the way of AP stacks allocation for the RLPs. TBOOT calculates the AP stack offset, basing on
the X2 APIC ID's value of the given RLP. When the APIC ID exceeded the NR_CPUS (1024), then
the RLP wakeup process execution halted. For the current moment, we increased the maximal X2 APID value
from 1024 to 8192. It is the similar value to this, which was defined in the Linux Kernel code.
This change fixes described issue.

In the future, we are going to replace the current form of AP stack
allocation with some other method e.g: dynamic stack allocation for each CPU core or by using OsMleData
stuct for this purpose.