Eliminate compilation errors, caused by the unsupported movl and pushl CPU instructions.
The compilation process of the TBOOT project has failed, when it was built and run
with the -m64 and the target_arch set as x86-64. GNU compiler couldn't recognize movl
and pushl instructions, used in the project. It was necessary to remove them and update
some implementations of the mov instructions, that they matched x86-64 architecture requirements.

It is still necessary to get rid of the compilation errors, appearing in the GDT table and
NMI interrupts configuration procedures.