Increase DIRECTMAP size from 64 MB to 128 MB
The DIRECTMAP is a memory map used to map TBOOT MLE, Serial log and Kernel
command pages. It is created to verify memory integrity, during the TBOOT
shutdown. During the last TBOOT validation on the Intel ARL-S CPUs, it was
noticed that TBOOT returns FATAL error, because TBOOT pages exceeded the maximal
memory size, which can be mapped by the DIRECTMAP.